fix showinfor
